move smoothMixel L21 Norm test in test_functions
